Fix ShippingService numeric type mismatches and CartController nullable guard

+++ Services/ShippingModels.cs
@@ -2,11 +2,11 @@ namespace Datn.PcStore.Services;
 
 public record GeoPoint(double Latitude, double Longitude);
 
-public record RouteMetrics(double DistanceKm, double DurationMinutes);
+public record RouteMetrics(decimal DistanceKm, int DurationMinutes);
 
 public class ShippingQuote
 {
-    public double DistanceKm { get; set; }
+    public decimal DistanceKm { get; set; }
     public int DurationMinutes { get; set; }
     public decimal ShippingFee { get; set; }
     public string Provider { get; set; } = string.Empty;
